Barcode: move exceptions in sub namespace Exception
