+++ library/Zend/Barcode/Exception/InvalidArgumentException.php
@@ -22,7 +22,8 @@
 /**
  * @namespace
  */
-namespace Zend\Barcode;
+namespace Zend\Barcode\Exception;
+use Zend\Barcode\Exception;
 
 /**
  * Exception for Zend_Barcode component.
@@ -33,7 +34,7 @@ namespace Zend\Barcode;
  * @copyright  Copyright (c) 2005-2010 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
-class RendererCreationException
+class InvalidArgumentException
     extends \InvalidArgumentException
     implements Exception
 {
