+++ library/Zend/Barcode/Object/Exception/OutOfRangeException.php
@@ -22,7 +22,8 @@
 /**
  * @namespace
  */
-namespace Zend\Barcode\Renderer;
+namespace Zend\Barcode\Object\Exception;
+use Zend\Barcode\Object\Exception;
 
 /**
  * Exception for Zend_Barcode component.
@@ -33,8 +34,8 @@ namespace Zend\Barcode\Renderer;
  * @copyright  Copyright (c) 2005-2010 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
-class RuntimeException
-    extends \RuntimeException
+class OutOfRangeException
+    extends \OutOfRangeException
     implements Exception
 {
 }
