+++ library/Zend/Barcode/Object/Exception/RuntimeException.php
@@ -22,7 +22,8 @@
 /**
  * @namespace
  */
-namespace Zend\Barcode\Renderer;
+namespace Zend\Barcode\Object\Exception;
+use Zend\Barcode\Object\Exception;
 
 /**
  * Exception for Zend_Barcode component.
@@ -33,8 +34,8 @@ namespace Zend\Barcode\Renderer;
  * @copyright  Copyright (c) 2005-2010 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
-class OutOfRangeException
-    extends \OutOfRangeException
+class RuntimeException
+    extends \RuntimeException
     implements Exception
 {
 }
