+++ library/Zend/Barcode/Renderer/Exception/RuntimeException.php
@@ -22,7 +22,8 @@
 /**
  * @namespace
  */
-namespace Zend\Barcode\Object;
+namespace Zend\Barcode\Renderer\Exception;
+use Zend\Barcode\Renderer\Exception;
 
 /**
  * Exception for Zend_Barcode component.
