+++ library/Zend/Barcode/Renderer/Exception/UnexpectedValueException.php
@@ -22,7 +22,8 @@
 /**
  * @namespace
  */
-namespace Zend\Barcode\Renderer;
+namespace Zend\Barcode\Renderer\Exception;
+use Zend\Barcode\Renderer\Exception;
 
 /**
  * Exception for Zend_Barcode component.
