+++ library/Zend/Barcode/Renderer/Image.php
@@ -24,7 +24,10 @@
  * @namespace
  */
 namespace Zend\Barcode\Renderer;
-use Zend\Barcode;
+use Zend\Barcode,
+    Zend\Barcode\Renderer\Exception\RuntimeException,
+    Zend\Barcode\Renderer\Exception\OutOfRangeException,
+    Zend\Barcode\Renderer\Exception\InvalidArgumentException;
 
 /**
  * Class for rendering the barcode as image
